requirements.txt
Django==1.10
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.23
celery==3.1.18
distribute==0.6.24
djangorestframework==3.4.4
kombu==3.0.35
pytz==2016.6.1
redis==2.10.3
vine==1.1.1
wsgiref==0.1.2
